loooong

+++ ClientAgent/Components/substration/Sub.cs
@@ -26,8 +26,8 @@ namespace Sub
         public Sub()
         {
             this.ComponentGuid = Guid.NewGuid();
-            this.InputHints = new ReadOnlyCollection<string>(new[] { typeof(int).ToString(), typeof(int).ToString() });
-            this.OutputHints = new ReadOnlyCollection<string>(new[] { typeof(int).ToString() });
+            this.InputHints = new ReadOnlyCollection<string>(new[] { typeof(long).ToString(), typeof(long).ToString() });
+            this.OutputHints = new ReadOnlyCollection<string>(new[] { typeof(long).ToString() });
             this.InputDescriptions = new List<string>(new[] { "num1", "num2" });
             this.OutputDescriptions = new List<string>(new[] { "result" });
         }
@@ -105,11 +105,11 @@ namespace Sub
                 yield return new object[] { new ArgumentException() };
             }
 
-            int result = 0;
+            long result = 0;
 
             foreach (var item in values)
             {
-                result -= (int)item;
+                result -= (long)item;
             }
 
             yield return result;
